refactor(kda): separate prefill backend from decode API
