+++ benchmarks/bench_recurrent_kda_prefill.py
@@ -42,10 +42,8 @@ from typing import Callable, Optional
 import numpy as np
 import torch
 
-from flashinfer.kda import (
-    RecurrentKDAPrefillWorkspace,
-    recurrent_kda,
-)
+from flashinfer.kda_decode import recurrent_kda
+from flashinfer.kda_prefill import RecurrentKDAPrefillWorkspace
 from flashinfer.testing import bench_gpu_time
 from flashinfer.utils import get_compute_capability
