+++ docs/index.rst
@@ -53,7 +53,8 @@ FlashInfer is a library and kernel generator for Large Language Models that prov
    api/activation
    api/gdn_decode
    api/gdn_prefill
-   api/kda
+   api/kda_decode
+   api/kda_prefill
    api/mamba
    api/mhc
    api/quantization
